requirements.txt
mlflow>=2.16.2
cloudpickle==2.2.1
configparser==5.2.0
cryptography==41.0.3
databricks-agents==0.5.0
databricks-vectorsearch==0.40
google-cloud-storage==2.10.0
grpcio-status==1.60.0
langchain-community==0.2.4
langchain==0.2.1
numpy==1.23.5
pandas==1.5.3
pyarrow==14.0.1
pydantic==1.10.6
pyspark==3.5.0